api notifications/dismiss -> ept-notifications
remove server/routes/notifications
