--- server/routes/notifications/index.js
@@ -1,7 +1,0 @@
-var path = require('path');
-var notifications = require(path.normalize(__dirname + '/config'));
-
-// Export Routes/Pre
-module.exports = [
-  { method: 'POST', path: '/notifications/dismiss', config: notifications.dismiss }
-];
